[Fix] SafeZone using dfs
