+++ seonhak/210704-안전영역, 구간합구하기5, 동전1, 단지번호붙이기/SafeZoneDfs.java
@@ -6,6 +6,7 @@ import java.io.IOException;
 import java.io.InputStreamReader;
 import java.io.OutputStreamWriter;
 import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.Collections;
 import java.util.List;
 
@@ -17,54 +18,53 @@ import java.util.List;
  */
 public class SafeZoneDfs {
 
-	static int complexMapSize = 0; //단지 지도 크기
-	static int homeCnt = 0;//각 단지 집수
-	static boolean[] DFSisVisited;
-	static List<Integer> complexList = new ArrayList<Integer>(); //단지 크기 정보
-	static int limitNum=0;
+	static int[][] safeZoneMap;//안전구역 정보
+	static int safeZoneMapSize = 0; //안전구역 맵 크기
+	static boolean[][] DFSisVisited;
+	static List<Integer> list;//각 수위 경계값에 따른 결과 값 저장
+	static int[] dx = {1, -1, 0, 0}; //상하좌우위아래
+	static int[] dy = {0, 0, 1, -1}; 
 	public static void main(String[] args) throws IOException {
 
 		BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
 		BufferedWriter bw = new BufferedWriter(new OutputStreamWriter(System.out));
 
-		complexMapSize = Integer.valueOf(br.readLine()); // 행렬 가로 길이
-		int[][] complexGraph = new int[complexMapSize + 1][complexMapSize + 1];
-		int[][] tempGraph = new int[complexMapSize + 1][complexMapSize + 1];
-
-		int max = Integer.MIN_VALUE;
-		int min = Integer.MAX_VALUE;
-		for (int i = 0; i < complexMapSize; i++) {
+		safeZoneMapSize = Integer.valueOf(br.readLine()); // 행렬 가로 길이
+		safeZoneMap = new int[safeZoneMapSize][safeZoneMapSize];
+		DFSisVisited = new boolean[safeZoneMapSize][safeZoneMapSize];
+		
+		int max = 0;
+		for (int i = 0; i < safeZoneMapSize; i++) {
 			String[] complexGraphInfo = String.valueOf(br.readLine()).split(" ");
-			for (int j = 0; j < complexMapSize; j++) {
-				complexGraph[i][j] = Integer.valueOf(complexGraphInfo[j]);
-				if(Integer.valueOf(complexGraphInfo[j])>=max)
-					max = Integer.valueOf(complexGraphInfo[j]);
-				if(Integer.valueOf(complexGraphInfo[j])<=min)
-					min = Integer.valueOf(complexGraphInfo[j]);
+			for (int j = 0; j < safeZoneMapSize; j++) {
+				safeZoneMap[i][j] = Integer.valueOf(complexGraphInfo[j]);
+				if(max < safeZoneMap[i][j])
+					max = safeZoneMap[i][j];
 			}
 		}
-		System.out.println("max : "+max);
-		System.out.println("min : "+min);
-		tempGraph = complexGraph;
-		int result = Integer.MIN_VALUE;
-		for(int k=max; k >= min; k--) {	
-			limitNum = k;
-			homeCnt = 0;
-			complexGraph = tempGraph;
-				for (int j = 0; j < complexMapSize; j++) {
-					for (int i = 0; i < complexMapSize; i++) {
-						if (complexGraph[i][j] > limitNum) {
-							homeCnt++;
-							dfs(complexGraph, i, j);
-						}
+		
+		list = new ArrayList<>();
+		
+		for(int depth=0; depth<=max ; depth++) {
+			int cnt = 0; 
+			for (int i = 0; i < safeZoneMapSize; i++) {
+				for (int j = 0; j < safeZoneMapSize; j++) {
+					if(safeZoneMap[i][j] > depth && !DFSisVisited[i][j]) {
+						cnt++;
+						dfs(i, j, depth);
 					}
+					
 				}
-				System.out.println(homeCnt);
-				if(result <= homeCnt)
-					result = homeCnt;
+			}
+			
+			//dfs 방문정보 초기화!!!!
+			for(boolean a[]:DFSisVisited)
+				Arrays.fill(a, false);
+			list.add(cnt);
 		}
 		
-		bw.write(String.valueOf(result) );
+		int maxResult = Collections.max(list);
+		bw.write(String.valueOf(maxResult) );
 
 		br.close();
 		bw.flush();
@@ -72,19 +72,20 @@ public class SafeZoneDfs {
 
 	}
 
-	static void dfs(int[][] graph, int i, int j) {
-		int m = complexMapSize, n = complexMapSize;
+	public static void dfs(int x, int y, int depth) {
+		DFSisVisited[x][y] = true;
+		
+		for(int i=0; i < 4; i++) {
+			int nx = x + dx[i];
+			int ny = y + dy[i];
+			
+			if(nx >=0 && ny >=0 && nx < safeZoneMap.length && ny< safeZoneMap.length) {
+				if(safeZoneMap[nx][ny] > depth && !DFSisVisited[nx][ny])
+					dfs(nx, ny, depth);
+			}
+			
+		}
 
-		if (i < 0 || i >= m || j < 0 || j >= n || graph[i][j] <= limitNum)
-			return;
-
-		//첫 단지 시작 이외 접근한 집은 -1처리
-		graph[i][j] = -1;
-
-		dfs(graph, i - 1, j);// 위
-		dfs(graph, i + 1, j);// 아래
-		dfs(graph, i, j - 1);// 왼쪽
-		dfs(graph, i, j + 1);// 오른쪽
 	}
 
 }
