+++ seonhak/210704-안전영역, 구간합구하기5, 동전1, 단지번호붙이기/SafeZoneDfsTry.java
@@ -1,0 +1,90 @@
+package com.ji.beakjoon.dfs;
+
+import java.io.BufferedReader;
+import java.io.BufferedWriter;
+import java.io.IOException;
+import java.io.InputStreamReader;
+import java.io.OutputStreamWriter;
+import java.util.ArrayList;
+import java.util.Collections;
+import java.util.List;
+
+/**
+ * 안전 영역 
+ * (실패)
+ * @author ji
+ *
+ */
+public class SafeZoneDfs {
+
+	static int complexMapSize = 0; //단지 지도 크기
+	static int homeCnt = 0;//각 단지 집수
+	static boolean[] DFSisVisited;
+	static List<Integer> complexList = new ArrayList<Integer>(); //단지 크기 정보
+	static int limitNum=0;
+	public static void main(String[] args) throws IOException {
+
+		BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
+		BufferedWriter bw = new BufferedWriter(new OutputStreamWriter(System.out));
+
+		complexMapSize = Integer.valueOf(br.readLine()); // 행렬 가로 길이
+		int[][] complexGraph = new int[complexMapSize + 1][complexMapSize + 1];
+		int[][] tempGraph = new int[complexMapSize + 1][complexMapSize + 1];
+
+		int max = Integer.MIN_VALUE;
+		int min = Integer.MAX_VALUE;
+		for (int i = 0; i < complexMapSize; i++) {
+			String[] complexGraphInfo = String.valueOf(br.readLine()).split(" ");
+			for (int j = 0; j < complexMapSize; j++) {
+				complexGraph[i][j] = Integer.valueOf(complexGraphInfo[j]);
+				if(Integer.valueOf(complexGraphInfo[j])>=max)
+					max = Integer.valueOf(complexGraphInfo[j]);
+				if(Integer.valueOf(complexGraphInfo[j])<=min)
+					min = Integer.valueOf(complexGraphInfo[j]);
+			}
+		}
+		System.out.println("max : "+max);
+		System.out.println("min : "+min);
+		tempGraph = complexGraph;
+		int result = Integer.MIN_VALUE;
+		for(int k=max; k >= min; k--) {	
+			limitNum = k;
+			homeCnt = 0;
+			complexGraph = tempGraph;
+				for (int j = 0; j < complexMapSize; j++) {
+					for (int i = 0; i < complexMapSize; i++) {
+						if (complexGraph[i][j] > limitNum) {
+							homeCnt++;
+							dfs(complexGraph, i, j);
+						}
+					}
+				}
+				System.out.println(homeCnt);
+				if(result <= homeCnt)
+					result = homeCnt;
+		}
+		
+		bw.write(String.valueOf(result) );
+
+		br.close();
+		bw.flush();
+		bw.close();
+
+	}
+
+	static void dfs(int[][] graph, int i, int j) {
+		int m = complexMapSize, n = complexMapSize;
+
+		if (i < 0 || i >= m || j < 0 || j >= n || graph[i][j] <= limitNum)
+			return;
+
+		//첫 단지 시작 이외 접근한 집은 -1처리
+		graph[i][j] = -1;
+
+		dfs(graph, i - 1, j);// 위
+		dfs(graph, i + 1, j);// 아래
+		dfs(graph, i, j - 1);// 왼쪽
+		dfs(graph, i, j + 1);// 오른쪽
+	}
+
+}
